+++ langchain4j-alloydb-pg/src/main/java/dev/langchain4j/store/alloydb/engine/MetadataColumn.java
@@ -1,4 +1,4 @@
-package dev.langchain4j.store.embedding.alloydb;
+package dev.langchain4j.store.alloydb.engine;
 
 public class MetadataColumn {
